Copy all relevant files from Efficient-CapsNet-Thesis
Initially, I wanted to build on the ideas presented in Efficient-CapsNet. Ultimately, I diverged from that some much that I do not use any of their classes anymore, or have changed them so much that it is no longer compatible with their models. To keep my work easier to understand I decided to move it to a new repo, without all of the files I do not use.