Pass step through the optimizer
By passing the steps concatened with the activations, I was messing up gradients. Luckily I found that the optimizer tracks the steps, and the optimizer could be passed to the loss functions. So now, the loss function checks the steps through the optimizer. Also, the paper claims that Spread Loss is equivalent to Squared Hinge Loss if m=1, but it really is not. In this set-up, I made it so that SpreadLoss IS equivalent to SquaredHinge. Will need some further research to see what works best:)